feat 닷엔브 설정 && 컴포넌트 분리
