+++ .env
@@ -1,0 +1,2 @@
+REST_API_KEY = 5f9dd5b73e2f948b6654fcc79a476b5a
+REDIRECT_URI = http://172.30.1.1:8081/Home
