+++ babel.config.js
@@ -1,6 +1,19 @@
-module.exports = function(api) {
-  api.cache(true);
-  return {
-    presets: ['babel-preset-expo'],
-  };
+module.exports = function (api) {
+    api.cache(true);
+    return {
+        presets: ["babel-preset-expo"],
+        plugins: [
+            [
+                "module:react-native-dotenv",
+                {
+                    moduleName: "@env",
+                    path: ".env",
+                    blacklist: null,
+                    whitelist: null,
+                    safe: false,
+                    allowUndefined: true,
+                },
+            ],
+        ],
+    };
 };
