+++ env.d.ts
@@ -1,0 +1,4 @@
+declare module "@env" {
+    export const REST_API_KEY: string;
+    export const REDIRECT_URI: string;
+}
