+++ package.json
@@ -17,6 +17,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.73.4",
+    "react-native-dotenv": "^3.4.11",
     "react-native-gesture-handler": "^2.15.0",
     "react-native-safe-area-context": "^4.9.0",
     "react-native-web": "~0.19.6",
@@ -25,6 +26,7 @@
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.45",
+    "@types/react-native-dotenv": "^0.2.2",
     "husky": "^9.0.11",
     "typescript": "^5.1.3"
   },
