+++ src/components/Auth/KakaoLogin.tsx
@@ -1,0 +1,37 @@
+import React from "react";
+import { REST_API_KEY, REDIRECT_URI } from "@env";
+import WebView from "react-native-webview";
+
+const KakaoLogin = () => {
+    const getCode = (target: string) => {
+        const exp = "code=";
+        const condition = target.indexOf(exp);
+        if (condition !== -1) {
+            const requestCode = target.substring(condition + exp.length);
+            console.log("code = ", requestCode);
+        }
+    };
+
+    const handleNavigationStateChange = (navState: {
+        url: string | string[];
+    }) => {
+        if (navState.url.includes(REDIRECT_URI)) {
+        }
+    };
+
+    return (
+        <WebView
+            style={{ flex: 1 }}
+            source={{
+                uri: `https://kauth.kakao.com/oauth/authorize?response_type=code&client_id=${REST_API_KEY}&redirect_uri=${REDIRECT_URI}`,
+            }}
+            onMessage={(event) => {
+                const data = event.nativeEvent.url;
+                getCode(data);
+            }}
+            onNavigationStateChange={handleNavigationStateChange}
+        />
+    );
+};
+
+export default KakaoLogin;
